apps/ml-engine/pyproject.toml
[project]
name = "ml-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "google-generativeai (>=0.8.6,<0.9.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
